- introduced pre-calculated tables for integer decoders run-time table initialization which includes many expensive FP operations is a very heavy task for the CPUs without FPU (sometimes takes more cpu time than decoding itself, the cause of stuttering at startup). the binary size increases(about +32KB), but significantly faster on those CPUs. undef PRECALC_TABLES macro in mpg123lib_intern.h to suppress this behavior. - faster decwin table initialization for integer decoders we can't adopt pre-calculation for this table due to RVA, so replaced some floating-point maths with fixed-point ones for faster initialization. - removed COS1 table from layer3.c, which is no longer used anywhere.
git-svn-id: svn://scm.orgis.org/mpg123/trunk@2159 35dc7657-300d-0410-a2e5-dc2837fedb53

+++ src/libmpg123/decode.h
@@ -213,8 +213,8 @@ void init_layer3_stuff(mpg123_handle *fr, real (*gainpow2)(mpg123_handle *fr, in
 #endif
 #ifndef NO_LAYER12
 void  init_layer12(void);
-real* init_layer12_table(mpg123_handle *fr, real *table, double m);
-void  init_layer12_stuff(mpg123_handle *fr, real* (*init_table)(mpg123_handle *fr, real *table, double m));
+real* init_layer12_table(mpg123_handle *fr, real *table, int m);
+void  init_layer12_stuff(mpg123_handle *fr, real* (*init_table)(mpg123_handle *fr, real *table, int m));
 #endif
 
 void prepare_decode_tables(void);
@@ -232,7 +232,7 @@ void make_decode_tables_mmx(mpg123_handle *fr);
 real init_layer3_gainpow2_mmx(mpg123_handle *fr, int i);
 #endif
 #ifndef NO_LAYER12
-real* init_layer12_table_mmx(mpg123_handle *fr, real *table, double m);
+real* init_layer12_table_mmx(mpg123_handle *fr, real *table, int m);
 #endif
 #endif
